✨ ⚡️Change PUT to PATCH in all resource APIs and minor updates

+++ dataservice/api/participant/resources.py
@@ -1,6 +1,4 @@
-from flask import abort, request, current_app
-from flask.views import MethodView
-from sqlalchemy.orm.exc import NoResultFound
+from flask import abort, request
 from sqlalchemy.orm import joinedload
 from marshmallow import ValidationError
 
@@ -83,16 +81,16 @@ class ParticipantAPI(CRUDView):
             resource:
               Participant
         """
-        try:
-            participant = Participant.query.filter_by(kf_id=kf_id).one()
-        except NoResultFound:
+        p = Participant.query.get(kf_id)
+        if p is None:
             abort(404, 'could not find {} `{}`'
-                  .format('Participant', kf_id))
-        return ParticipantSchema().jsonify(participant)
+                  .format('participant', kf_id))
 
-    def put(self, kf_id):
+        return ParticipantSchema().jsonify(p)
+
+    def patch(self, kf_id):
         """
-        Update an existing participant
+        Update an existing participant. Allows partial update
         ---
         template:
           path:
@@ -101,23 +99,25 @@ class ParticipantAPI(CRUDView):
             resource:
               Participant
         """
-        body = request.json
-        try:
-            p = Participant.query.filter_by(kf_id=kf_id).one()
-        except NoResultFound:
+        p = Participant.query.get(kf_id)
+        if p is None:
             abort(404, 'could not find {} `{}`'
-                  .format('Participant', kf_id))
+                  .format('participant', kf_id))
 
-        p.external_id = body.get('external_id')
-        p.family_id = body.get('family_id')
-        p.is_proband = body.get('is_proband')
-        p.consent_type = body.get('consent_type')
-        p.study_id = body.get('study_id')
+        # Partial update - validate but allow missing required fields
+        body = request.json or {}
+        try:
+            p = ParticipantSchema(strict=True).load(body, instance=p,
+                                                    partial=True).data
+        except ValidationError as err:
+            abort(400, 'could not update participant: {}'.format(err.messages))
+
+        db.session.add(p)
         db.session.commit()
 
         return ParticipantSchema(
-            201, 'participant {} updated'.format(p.kf_id)
-        ).jsonify(p), 201
+            200, 'participant {} updated'.format(p.kf_id)
+        ).jsonify(p), 200
 
     def delete(self, kf_id):
         """
@@ -130,10 +130,10 @@ class ParticipantAPI(CRUDView):
             resource:
               Participant
         """
-        try:
-            p = Participant.query.filter_by(kf_id=kf_id).one()
-        except NoResultFound:
-            abort(404, 'could not find {} `{}`'.format('Participant', kf_id))
+        p = Participant.query.get(kf_id)
+        if p is None:
+            abort(404, 'could not find {} `{}`'
+                  .format('participant', kf_id))
 
         db.session.delete(p)
         db.session.commit()
